requirements.txt
ollama
requests
ultralytics
opencv-python
pyserial
speechrecognition
pyttsx3
pyaudio
pyyaml
